+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog of threedi-schema
 0.224.7 (unreleased)
 --------------------
 
+- Prevent migration 222 to 224 from failing when any of the new table names already exists
 - Swap start and end of control_measure_map geometries
 - Modify geometry of controls associated with pumpstation to the pumpstation start node
 - Ensure control_measure_map.geom is a valid line